% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars2.R
\docType{data}
\name{cars2}
\alias{cars2}
\title{cars2}
\format{
A data frame with 50 rows and 4 variables:
\describe{
\item{high_speed}{Logical indicating whether speed is higher than the average}
\item{speed}{After original cars speed}
\item{dist}{After original dist speed}
\item{high_dist}{Logical indicating whether dist is higher than the average}
}
}
\source{
dataset package
}
\usage{
cars2
}
\description{
An adaptation after the \code{cars} dataset which includes logical variables.
}
\keyword{datasets}
