% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_filter-utils.R
\name{col_vals}
\alias{col_vals}
\alias{col_vals.numeric}
\title{Base filter method based on the class of column}
\usage{
col_vals(col, remove_na = FALSE)

\method{col_vals}{numeric}(col, remove_na = FALSE)
}
\arguments{
\item{col}{Column}

\item{remove_na}{Remove NA's}
}
\value{
List containing the ingredients to build a Shiny controller.
}
\description{
Base filter method based on the class of column
}
