% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_metric}
\alias{create_metric}
\title{Helper function for shiny server module}
\usage{
create_metric(name, calculation, description = NULL)
}
\arguments{
\item{name}{a dataframe with ad campaign data}

\item{calculation}{a character string specifying the title of the table}

\item{description}{a character string specifying the title of the table}
}
\value{
a list
}
\description{
Helper function for shiny server module
}
\examples{
\donttest{
ametric <- create_metric("cpc", "sum(spend) / sum(clicks)", "Cost per click"
}
}
