% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-module.R
\name{dataset_ui}
\alias{dataset_ui}
\alias{dataset_server}
\title{Hadley's Dataset Module}
\usage{
dataset_ui(id, filter = NULL)

dataset_server(id)
}
\arguments{
\item{id}{Namespace of the module.}

\item{filter}{Filter the datasets (e.g. \code{is.data.frame}).}
}
\value{
Shiny module.
}
\description{
Hadley's Dataset Module
}
