% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-module.R
\name{download_ui}
\alias{download_ui}
\alias{download_server}
\title{Download module}
\usage{
download_ui(id, label = "Download", n)

download_server(id, data, name, n, sep = ",")
}
\arguments{
\item{id}{Namespace of the module.}

\item{label}{On download button.}

\item{n}{The number of download handles.}

\item{data}{The output data.}

\item{name}{Name of the output file.}

\item{sep}{Determines the type of file \verb{,} for a csv-file and \verb{\\t} for a tab-
delimited text file.}
}
\value{
Shiny module.
}
\description{
The module is constructed to allow creating multiple download buttons.
}
