% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_census_data}
\alias{get_census_data}
\title{Get Census Data}
\usage{
get_census_data(ethnicity)
}
\arguments{
\item{ethnicity}{Character string specifying the ethnicity to query}
}
\value{
sf object with county geometries and population data
}
\description{
Get Census Data
}
\examples{
\donttest {
library(tidycensus)
df <- get_census_data("white")
}
}
