% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataDemo.R
\name{getDataDemo}
\alias{getDataDemo}
\title{Example get imported data application}
\usage{
getDataDemo()
}
\description{
This is an application using the \code{get_flat_data()} function.
}
\details{
\subsection{App structure}{

The UI and server contains examples from the following \code{get_flat_data()}
function arguments:

\if{html}{\out{<div class="sourceCode">}}\preformatted{get_flat_data(name = 'import', id = 'data', mod = TRUE, return_data = TRUE)
}\if{html}{\out{</div>}}

\emph{and}

\if{html}{\out{<div class="sourceCode">}}\preformatted{get_flat_data(name = 'import', id = 'data', mod = FALSE, return_data = FALSE)
}\if{html}{\out{</div>}}
}

\subsection{Module function arguments}{

If \code{mod} and \code{return_data} are set to \code{TRUE}, then the module will not
display the imported data. To verify the data have been imported,
\code{getDataDemo()} has an additional module (\code{mod_prnt_data_ui()/server()}) that
renders the imported data via \code{shiny::verbatimTextOutput()} and
\code{shiny::renderPrint()}.
}

\subsection{UI/server arguments}{

If \code{mod} is set to \code{FALSE}, \code{get_flat_data()} returns UI and server code for
importing data and returning a reactive object (\code{imported_data()}).
\code{getDataDemo()} also displays these data using \code{shiny::verbatimTextOutput()}
and \code{shiny::renderPrint()}
}
}
