% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTS.R
\docType{data}
\name{GTS}
\alias{GTS}
\title{GTS}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 115 rows and 5 columns.
}
\usage{
GTS
}
\description{
Geological timescale from \href{paleobiodb}{https://paleobiodb.org/}.
}
\keyword{datasets}
