% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_flat_file.R
\name{import_flat_file}
\alias{import_flat_file}
\title{Import flat data files}
\usage{
import_flat_file(path)
}
\arguments{
\item{path}{path to data file (with extension)}
}
\value{
return_data as loaded flat file
}
\description{
Imports a variety of flat file types (SAS, SPSS, Stata, and
plain text files).
}
