% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLeafletMap.R
\name{leafletMapServer}
\alias{leafletMapServer}
\title{The leafletMapServer module to generate the map and interactivity}
\usage{
leafletMapServer(id, df)
}
\arguments{
\item{id}{namespace id}

\item{df}{data frame}
}
\value{
A shiny module server
}
\description{
The leafletMapServer module to generate the map and interactivity
}
\examples{
\donttest{
library(shiny)
ui <- fluidPage(pointMapUI("pointmaptest"))

server <- function(input, output, session) {
  pointMapServer("pointmaptest", df = reactive({ covid }))
}
shinyApp(ui, server)
}
}
