% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLeafletMap.R
\name{leafletMapUI}
\alias{leafletMapUI}
\title{The leafletMapUI module to generate the user interface}
\usage{
leafletMapUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
A shiny module UI
}
\description{
The leafletMapUI module to generate the user interface
}
\examples{
library(shiny)
ui <- fluidPage(pointMapUI("pointmaptest"))

server <- function(input, output, session) {
  pointMapServer("pointmaptest", df = reactive({ covid }))
}
shinyApp(ui, server)
}
