% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modMap.R
\name{mapUI}
\alias{mapUI}
\title{County Map Module UI}
\usage{
mapUI(id)
}
\arguments{
\item{id}{The module ID used to create a namespace}
}
\value{
A Shiny UI element containing the map interface
}
\description{
Creates the UI portion of the county map module for displaying
demographic data choropleth visualizations.
}
\examples{
library(shiny)
ui <- fluidPage(mapUI("map1"))

server <- function(input, output, session) {

 # mydata <- readRDS(here("data", "county_eth.RDS"))
 mapServer("map1", df = reactive({ county_pop }) )

 }
shinyApp(ui, server)
}
