% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_greet_input.R
\name{mod_greet_input_server}
\alias{mod_greet_input_server}
\title{Greeting application inputs (server)}
\usage{
mod_greet_input_server(id)
}
\arguments{
\item{id}{namespace'd id}
}
\description{
This is an example UI module function that combines text input
and an action button to create a simple greeting app. It's been adapted from
the \href{https://www.youtube.com/watch?v=Gucwz865aqQ}{\code{shinytest2} demo on YouTube}
}
