% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_greet_output.R
\name{mod_greet_output_server}
\alias{mod_greet_output_server}
\title{Greeting application output (server)}
\usage{
mod_greet_output_server(id, greeting)
}
\arguments{
\item{id}{namespace'd id}
}
\description{
This is an example UI module function that displays the text
output that's been collected from \code{mod_greet_input_ui/server()}. It's been
adapted from the \href{https://www.youtube.com/watch?v=Gucwz865aqQ}{\code{shinytest2} demo on YouTube}
}
