% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataDemo.R
\name{mod_import_server}
\alias{mod_import_server}
\title{Example get imported flat file data server function}
\usage{
mod_import_server(id)
}
\arguments{
\item{id}{}
}
\description{
This is an example server function from \code{get_flat_data()}
}
\details{
\subsection{Function arguments}{

The following arguments were used to generate \code{mod_import_ui/server()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{get_flat_data(name = "import", id = "data", mod = TRUE, return_data = TRUE)
}\if{html}{\out{</div>}}
}

\subsection{Return}{

\code{return_data = TRUE} will return the imported data as a reactive list.
}
}
