% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataDemo.R
\name{mod_prnt_data_server}
\alias{mod_prnt_data_server}
\title{Print the returned values from \code{get_flat_data()} (server)}
\usage{
mod_prnt_data_server(id, imported_flat_data)
}
\arguments{
\item{id}{}
}
\description{
This module is used in \code{getDataDemo()} to demonstrate how to
display the returned values from \code{get_flat_data()}.
}
