% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataDemo.R
\name{mod_prnt_data_ui}
\alias{mod_prnt_data_ui}
\title{Print the returned values from \code{get_flat_data()} (UI)}
\usage{
mod_prnt_data_ui(id)
}
\arguments{
\item{id}{}
}
\description{
This module is used in \code{getDataDemo()} to demonstrate how to
display the returned values from \code{get_flat_data()}.
}
