% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reactvals.R
\name{mod_reactvals_server}
\alias{mod_reactvals_server}
\title{Print reactive values (server)}
\usage{
mod_reactvals_server(id)
}
\arguments{
\item{id}{namespace'd id}
}
\description{
Print reactive values (server)
}
