% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactValsDemo.R
\name{mod_vals_ui}
\alias{mod_vals_ui}
\title{Example print reactive values UI function}
\usage{
mod_vals_ui(id)
}
\arguments{
\item{id}{}
}
\description{
This is an example of the \code{prnt_reactvals()} function.
}
\details{
\subsection{Function arguments}{

The following arguments were used to generate this module:

\if{html}{\out{<div class="sourceCode">}}\preformatted{prnt_reactvals(name = "vals", id = "dev", mod = TRUE)
}\if{html}{\out{</div>}}
}
}
