% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modTableBuilder.R
\name{modTblBuilderServer}
\alias{modTblBuilderServer}
\title{Create a server module for interacting with tabular data}
\usage{
modTblBuilderServer(id, df, filters, summary_fn)
}
\arguments{
\item{id}{the reference id for module}

\item{df}{a dataframe}

\item{filters}{stuff here}

\item{summary_fn}{your summary with metrics}
}
\value{
a UI module for tables
}
\description{
Create a server module for interacting with tabular data
}
\examples{
\donttest{
modTblBuilderServer()
}
}
