% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modTableBuilder.R
\name{modTblBuilderUI}
\alias{modTblBuilderUI}
\title{Create a user interface module for interacting with tabular data}
\usage{
modTblBuilderUI(id)
}
\arguments{
\item{id}{the reference id for module}
}
\value{
a UI module for tables
}
\description{
Create a user interface module for interacting with tabular data
}
\examples{
\donttest{
modTblBuilderUI()
}
}
