% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modPointMap.R
\name{pointMapServer}
\alias{pointMapServer}
\title{The pointMapServer module to generate the server logic}
\usage{
pointMapServer(id, df)
}
\arguments{
\item{id}{namespace id}

\item{df}{data frame}
}
\value{
A shiny module server
}
\description{
The pointMapServer module to generate the server logic
}
\examples{
ui <- fluidPage(pointMapUI("pointmaptest"))

server <- function(input, output, session) {
  pointMapServer("pointmaptest", df = reactive({ covid }))
}
shinyApp(ui, server)
}
