% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modPointMap.R
\name{pointMapUI}
\alias{pointMapUI}
\title{The pointMapUI module to generate the user interface}
\usage{
pointMapUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
A shiny module UI
}
\description{
The pointMapUI module to generate the user interface
}
\examples{
\donttest{
library(shiny)
ui <- fluidPage(pointMapUI("pointmaptest"))

server <- function(input, output, session) {
  pointMapServer("pointmaptest", df = reactive({ covid }))
}
shinyApp(ui, server)
}
}
