% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prnt_reactvals.R
\name{prnt_reactvals}
\alias{prnt_reactvals}
\title{Create a 'print reactive values' module}
\usage{
prnt_reactvals(name, id, mod = TRUE)
}
\arguments{
\item{name}{name of the module}

\item{id}{namespace id}
}
\value{
code for UI and server components
}
\description{
Create a 'print reactive values' module
}
\examples{
prnt_reactvals(name = "vals", id = "dev", mod = TRUE)
}
