% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prnt_rvals.R
\name{prnt_rvals}
\alias{prnt_rvals}
\title{Print reactive values}
\usage{
prnt_rvals(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
code for UI and server components
}
\description{
Print reactive values
}
\examples{
# prnt_rvals("dev")
}
