% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactValsDemo.R
\name{reactValsDemo}
\alias{reactValsDemo}
\title{Example print reactive values application}
\usage{
reactValsDemo()
}
\description{
This is an application using the \code{prnt_reactvals()} function.
}
\details{
\subsection{App structure}{

The UI and server contains \emph{both} outputs from the \code{prnt_reactvals()} function:

\if{html}{\out{<div class="sourceCode">}}\preformatted{prnt_reactvals(name = "vals", id = "dev", mod = TRUE)
}\if{html}{\out{</div>}}

\emph{and}

\if{html}{\out{<div class="sourceCode">}}\preformatted{prnt_reactvals(name = "vals", id = "dev", mod = FALSE)
}\if{html}{\out{</div>}}
}
}
