% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSales.R, R/modSalesBetter.R
\name{salesServer}
\alias{salesServer}
\title{The salesServer module to generate code.}
\usage{
salesServer(id, df, dropdowns, cols_for_table)

salesServer(id, df, dropdowns, cols_for_table)
}
\arguments{
\item{id}{namespace id}

\item{df}{a reactive dataframe with the data}

\item{dropdowns}{a list of three columns that will create the dropdown hiearchy}

\item{cols_for_table}{table columns to display in table}
}
\value{
A shiny module Server

A shiny module Server
}
\description{
The salesServer module to generate code.

The salesServer module to generate code.
}
\examples{
\donttest{
ui <- fluidPage(salesUI("sales1"))
server <- function(input, output, session) {
  salesServer("sales1", df = reactive({ sales }))
}
shinyApp(ui, server)
}
\donttest{
ui <- fluidPage(salesUI("sales1"))
server <- function(input, output, session) {
  salesServer("sales1",
              df = reactive({ sales }),
              dropdowns = list(
                level_1 = "TERRITORY",
                level_2 = "CUSTOMERNAME",
                level_3 = "ORDERNUMBER")
              ,
              cols_for_table = c("TERRITORY", "CUSTOMERNAME", "ORDERNUMBER", "SALES")
              )
}
shinyApp(ui, server)
ui <- fluidPage(salesUI("cars"))
server <- function(input, output, session) {
  salesServer("cars",
              df = reactive({ mtcars }),
              dropdowns = list(
                level_1 = "cyl",
                level_2 = "gear",
                level_3 = "carb")
              ,
              cols_for_table = c("mpg", "cyl", "disp", "hp", "wt")
  )
}
shinyApp(ui, server)
}

}
