% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSales.R
\name{salesUI}
\alias{salesUI}
\title{The salesUI module to generate code.}
\usage{
salesUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
A shiny module UI
}
\description{
The salesUI module to generate code.
}
\examples{
salesUI("sales1")

ui <- fluidPage(salesUI("sales1"))

server <- function(input, output, session) {

  salesServer("sales1", df = reactive({ sales }) )

}

shinyApp(ui, server)
}
