% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymods-package.R
\docType{package}
\name{shinymods-package}
\alias{shinymods}
\alias{shinymods-package}
\title{shinymods: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Martin Schobben \email{schobbenmartin@gmail.com} (\href{https://orcid.org/0000-0001-8560-0037}{ORCID})

}
\keyword{internal}
