% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendor.R
\name{shinymods_vendor}
\alias{shinymods_vendor}
\title{Vendor the shiny modul}
\usage{
shinymods_vendor(module, path = ".")
}
\arguments{
\item{module}{Character vector for the module (\code{"wizard"},
\code{"conditional_filter"})}

\item{path}{Path to your package (default is the current working directory).}
}
\value{
The file path to the vendored code (invisibly).
}
\description{
Vendoring makes the code available as a copy for you to use in your own
package.
}
