% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_metrics}
\alias{summarise_metrics}
\title{Basic summary statistics for ad campaign data}
\usage{
summarise_metrics(df, ...)
}
\arguments{
\item{df}{a dataframe with ad campaign data}

\item{...}{additional arguments to pass to \code{\link{summarise}}}
}
\value{
a dataframe with specific statistics summarising ad data
}
\description{
Basic summary statistics for ad campaign data
}
\examples{
\donttest{
my_summary <- summarise_metrics(df)
}
}
