% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_migration}
\alias{us_migration}
\title{A small us migration table downloaded with help from tidycensus}
\format{
\subsection{\code{us_migration}}{

A data frame with 52 rows and 7 columns
}
}
\usage{
us_migration
}
\description{
A small us migration table downloaded with help from tidycensus
}
\keyword{datasets}
