% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid}
\alias{covid}
\title{A covid dataset from Kaggle, used to test point mapping with the leaflet package}
\format{
\subsection{\code{covid}}{

A data frame with 20,706 rows and 11 columns
}
}
\source{
\url{https://www.kaggle.com/datasets/grebublin/coronavirus-latlon-dataset}
}
\usage{
covid
}
\description{
A covid dataset from Kaggle, used to test point mapping with the leaflet package
}
\keyword{datasets}
