% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_filter-utils.R
\name{detect_controller}
\alias{detect_controller}
\alias{logical_controller}
\alias{switch_controller}
\alias{col_vals.character}
\alias{col_vals.factor}
\alias{col_vals.logical}
\title{Build a controller based on data}
\usage{
detect_controller(
  dat,
  session,
  labels = character(1),
  external = character(1),
  remove_na = FALSE,
  update = FALSE
)

logical_controller(
  dat,
  session,
  labels = "Various",
  external = character(1),
  ignore = character(1),
  remove_na = FALSE,
  update = FALSE
)

switch_controller(col, x, method, val)

\method{col_vals}{character}(col, remove_na = FALSE)

\method{col_vals}{factor}(col, remove_na = FALSE)

\method{col_vals}{logical}(col, remove_na = FALSE)
}
\arguments{
\item{dat}{The data supplied as \code{dataframe} or tibble.}

\item{session}{Namespace of the module or session.}

\item{labels}{Labels for the controllers (defaults to the variable names).
Must be a named vector where the names refer to the names of the original
data.}

\item{external}{\code{reactiveValues} assigning external filter variables to the
module.}

\item{remove_na}{Logical indicating whether NAs should be removed during
filtering (defaults to FALSE).}

\item{update}{Logical indicating whether ui controllers are ui- or
server-side for updating the controllers ranges.}

\item{ignore}{Character string/vector for variables to be ignored for
filtering of the dataset.}
}
\value{
Shiny controller call.
}
\description{
Build a controller based on data
}
