% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_filter-module.R
\name{filter_ui}
\alias{filter_ui}
\alias{filter_server}
\title{Conditional filter module}
\usage{
filter_ui(
  dat,
  session,
  labels = character(1),
  external = character(1),
  logi = character(1),
  update = FALSE,
  shinyjs = FALSE,
  ignore = NULL,
  remove_na = FALSE
)

filter_server(
  id,
  dat,
  external = reactiveValues(),
  ignore = character(0),
  shinyjs = FALSE,
  remove_na = FALSE
)
}
\arguments{
\item{dat}{The data supplied as \code{dataframe} or tibble.}

\item{session}{The shiny session domain.}

\item{labels}{Labels for the controllers (defaults to the variable names).}

\item{external}{\code{reactiveValues} assigning external filter variables to the
module.}

\item{logi}{Label for the controller of logical variables (defaults to
\code{"Various"}.}

\item{update}{Logical indicating whether ui controllers are ui- or
server-side for updating the controllers ranges.}

\item{shinyjs}{A logical indicating whether shinyjs \code{ToggleState()} is
included to disable controllers.}

\item{ignore}{Character string/vector for variables to be ignored for
filtering of the dataset.}

\item{remove_na}{Logical indicating whether NAs should be removed during
filtering (defaults to FALSE).}

\item{id}{Namespace of the module.}
}
\value{
Shiny module.
}
\description{
The module is constructed so that the filter controls are updated based on
the data range.
}
