% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_filter-module.R
\name{filter_ui}
\alias{filter_ui}
\alias{filter_server}
\title{Conditional filter module}
\usage{
filter_ui(id)

filter_server(
  dat,
  id,
  labels = character(1),
  external = reactiveValues(),
  logi = character(1),
  ignore = character(1),
  shinyjs = FALSE,
  remove_na = FALSE
)
}
\arguments{
\item{id}{Namespace of the module.}

\item{dat}{The data supplied as \code{dataframe} or \code{tibble}.}

\item{labels}{Labels for the controllers (defaults to the variable names).}

\item{external}{\code{reactiveValues} assigning external filter variables to the
module.}

\item{logi}{Label for the controller of logical variables (defaults to
\code{"Various"}).}

\item{ignore}{Character string/vector for variables to be ignored for
filtering of the dataset.}

\item{shinyjs}{A logical indicating whether shinyjs \code{ToggleState()} is
included to disable controllers.}

\item{remove_na}{Logical indicating whether NAs should be removed during
filtering (defaults to FALSE).}
}
\value{
Shiny module.
}
\description{
The module is constructed so that the filter controls are updated based on
the data range.
}
