% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSales.R
\name{salesServer}
\alias{salesServer}
\title{The salesServer module to generate code.}
\usage{
salesServer(id, df)
}
\arguments{
\item{id}{namespace id}

\item{df}{a reactive dataframe with the data}
}
\value{
A shiny module Server
}
\description{
The salesServer module to generate code.
}
\examples{
salesServer("sales1")
}
