% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSales.R, R/modSalesBetter.R
\name{salesServer}
\alias{salesServer}
\title{The salesServer module to generate code.}
\usage{
salesServer(id, df, cols_list)

salesServer(id, df, cols_list)
}
\arguments{
\item{id}{namespace id}

\item{df}{a reactive dataframe with the data}

\item{cols_list}{a list of columns that will create the dropdowns selections}
}
\value{
A shiny module Server

A shiny module Server
}
\description{
The salesServer module to generate code.

The salesServer module to generate code.
}
