% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSales.R, R/modSalesBetter.R
\name{salesUI}
\alias{salesUI}
\title{The salesUI module to generate code.}
\usage{
salesUI(id)

salesUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
A shiny module UI

A shiny module UI
}
\description{
The salesUI module to generate code.

The salesUI module to generate code.
}
