% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSales.R, R/modSalesBetter.R
\name{salesUI}
\alias{salesUI}
\title{The salesUI module to generate code.}
\usage{
salesUI(id)

salesUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
A shiny module UI

A shiny module UI
}
\description{
The salesUI module to generate code.

The salesUI module to generate code.
}
\examples{
\donttest{
ui <- fluidPage(salesUI("sales1"))
server <- function(input, output, session) {
  salesServer("sales1", df = reactive({ sales }))
}
shinyApp(ui, server)
}
\donttest{
ui <- fluidPage(salesUI("sales1"))
server <- function(input, output, session) {
  salesServer("sales1",
              df = reactive({ sales }),
              dropdowns = list(
                level_1 = "TERRITORY",
                level_2 = "CUSTOMERNAME",
                level_3 = "ORDERNUMBER")
              ,
              cols_for_table = c("TERRITORY", "CUSTOMERNAME", "ORDERNUMBER", "SALES")
              )
}
shinyApp(ui, server)
}
}
